swagger and journary

+++ src/controllers/habits.js
@@ -4,15 +4,10 @@ const ObjectId = require('mongodb').ObjectId;
 const getAllHabits = async (req, res) => {
     try {
         //#swagger.tags=['Habits']
-        const result = await mongodb.getDb().collection('habitLogs').find();     
-        result.toArray().then((habits) => {        
-            res.setHeader('Content-Type', 'application/json');        
-            res.status(200).json(habits);      
-        })
-        .catch((err) => {          
-            res.status(500).json({ message: err.message });    
-        });
-
+        const result = await mongodb.getDb().collection('habitLogs').find();
+        const habits = await result.toArray();
+        res.setHeader('Content-Type', 'application/json');
+        res.status(200).json(habits);
     } catch (err) {   
         res.status(500).json({ message: err.message });  
     }
@@ -20,22 +15,21 @@ const getAllHabits = async (req, res) => {
 
 const getSingle = async (req, res) => {
     try {
+        if (!ObjectId.isValid(req.params.id)) {
+            return res.status(400).json({ message: 'Invalid habit id' });
+        }
         //#swagger.tags=['Habits']
         const habitId = new ObjectId(req.params.id);
         const result = await mongodb.getDb().collection('habitLogs').find({ _id: habitId });
-        result.toArray().then((habits) => {
-             if (!habits.length) {
-                return res.status(404).json({ message: 'Habit not found' });    
-            }
+        const habits = await result.toArray();
+        if (!habits.length) {
+            return res.status(404).json({ message: 'Habit not found' });    
+        }
       
-            res.setHeader('Content-Type', 'application/json');     
-            res.status(200).json(habits);     
-        })
-        .catch((err) => {      
-            res.status(500).json({ message: err.message });   
-        });
+        res.setHeader('Content-Type', 'application/json');     
+        res.status(200).json(habits);     
     } catch (err) { 
-        res.status(400).json({ message: 'Invalid habit id' });
+        res.status(500).json({ message: err.message });
     }
 };
 
@@ -56,6 +50,9 @@ const createHabit = async (req, res) => {
 
 const updateHabit = async (req, res) => {
     try {
+        if (!ObjectId.isValid(req.params.id)) {
+            return res.status(400).json({ message: 'Invalid habit id' });
+        }
         //#swagger.tags=['Habits']
         const habitId = new ObjectId(req.params.id);
         const habit = req.body;
@@ -66,22 +63,25 @@ const updateHabit = async (req, res) => {
             res.status(404).json({ message: 'Habit not found' });
         }
     } catch (err) {
-        res.status(400).json({ message: 'Invalid habit id' });
+        res.status(500).json({ message: err.message });
     }
 };
 
 const deleteHabit = async (req, res) => {
     try {
+        if (!ObjectId.isValid(req.params.id)) {
+            return res.status(400).json({ message: 'Invalid habit id' });
+        }
         //#swagger.tags=['Habits']
         const habitId = new ObjectId(req.params.id);
         const response = await mongodb.getDb().collection('habitLogs').deleteOne({ _id: habitId });
         if (response.deletedCount > 0) {
-        res.status(204).send();
+            res.status(204).send();
         } else {
-        res.status(404).json({ message: 'Habit not found' });
+            res.status(404).json({ message: 'Habit not found' });
         }
     } catch (err) {
-        res.status(400).json({ message: 'Invalid habit id' });
+        res.status(500).json({ message: err.message });
     }
 };
